local strage added

+++ src/app/components/logs/logs.component.ts
@@ -9,12 +9,24 @@ import {LogService} from '../../services/log.service';
 })
 export class LogsComponent implements OnInit {
   logs: Log[];
+  msg:boolean=false;
   constructor(private logService: LogService) {}
 
   ngOnInit(): void {
+    console.log('-----------on init', this.logs);
+    
     this.logService.getLogs().subscribe((logs) => {
       this.logs = logs;
+      
     });
+    console.log('--len--', this.msg);
+
+    this.logs.length == 0 || this.logs == null || this.logs == undefined
+      ? (this.msg = true)
+      : (this.msg = false); 
+
+      console.log('msg');
+      
   }
 
   onSelect(log: Log) {
@@ -23,7 +35,12 @@ export class LogsComponent implements OnInit {
   }
 
   onDelete(log:Log){
-    this.logService.deleteLog(log);
+    if(confirm('Are you sure, you want to delete?')){
+      this.logService.deleteLog(log);
+    }
 
   }
+
+
+  
 }
